Moved column pyramid series into subfolder.
