+++ js/masters/highcharts-more.src.js
@@ -13,7 +13,7 @@ import '../Core/Axis/RadialAxis.js';
 import '../Series/AreaRangeSeries.js';
 import '../Series/AreaSplineRangeSeries.js';
 import '../Series/ColumnRangeSeries.js';
-import '../Series/ColumnPyramidSeries.js';
+import '../Series/ColumnPyramid/ColumnPyramidSeries.js';
 import '../Series/GaugeSeries.js';
 import '../Series/BoxPlotSeries.js';
 import '../Series/ErrorBarSeries.js';
